Move const_ref_caster test to builtin_casters
